# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the core API for Sispeg Project in Azmeela",
        "title": "Sispeg API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Logs in a user using username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Logs out a user and revokes their refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout user",
                "parameters": [
                    {
                        "description": "Refresh Token Request",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Get a new access token using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "Refresh Token Request",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cash-advances": {
            "get": {
                "description": "Get a list of cash advance requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cash-advances"
                ],
                "summary": "Get cash advances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.CashAdvance"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Request a new cash advance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cash-advances"
                ],
                "summary": "Request cash advance",
                "parameters": [
                    {
                        "description": "Cash advance request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.CashAdvanceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.CashAdvance"
                        }
                    }
                }
            }
        },
        "/cash-advances/payment": {
            "post": {
                "description": "Record a payment toward a cash advance (debt)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cash-advances"
                ],
                "summary": "Add cash advance payment",
                "parameters": [
                    {
                        "description": "Payment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.CashAdvancePayment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cash-advances/{id}/status": {
            "put": {
                "description": "Approve or reject a cash advance request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cash-advances"
                ],
                "summary": "Update cash advance status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cash Advance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.CashAdvanceStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_azmeela_sispeg-api_internal_domain.AuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.CashAdvance": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "purpose": {
                    "type": "string"
                },
                "status": {
                    "description": "0: created, 1: approved, 2: disapproved",
                    "type": "integer"
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.CashAdvancePayment": {
            "type": "object",
            "required": [
                "amount",
                "date",
                "employee_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.CashAdvanceRequest": {
            "type": "object",
            "required": [
                "amount",
                "date",
                "purpose"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.CashAdvanceStatusUpdate": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.Employee": {
            "type": "object",
            "properties": {
                "base_salary": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "contract_end": {
                    "type": "string"
                },
                "contract_start": {
                    "type": "string"
                },
                "cv": {
                    "type": "string"
                },
                "date_sign": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_azmeela_sispeg-api_internal_domain.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_azmeela_sispeg-api_internal_domain.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        }
    }
}