# internal/delivery/http/dto/auth_dto.go
package dto

import (
	"github.com/azmeela/sispeg-api/internal/domain"
)

type LoginRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type AuthResponse struct {
	AccessToken  string            `json:"access_token"`
	RefreshToken string            `json:"refresh_token"`
	User         *EmployeeResponse `json:"user"`
}

type RefreshRequest struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}

func ToAuthResponse(accessToken, refreshToken string, user *domain.Employee) *AuthResponse {
	return &AuthResponse{
		AccessToken:  accessToken,
		RefreshToken: refreshToken,
		User:         ToEmployeeResponse(user),
	}
}